+++ h2o-algos/src/main/java/hex/DataInfo.java
@@ -8,6 +8,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import static water.util.ArrayUtils.findLongestCommonPrefix;
+
 /**
  * Created by tomasnykodym on 1/29/15.
  *
@@ -876,28 +878,6 @@ public class DataInfo extends Keyed<DataInfo> {
       j++;
     }
     return originalNames;
-  }
-  
-  public static String findLongestCommonPrefix(String inputArray[]) {
-    String referenceWord = inputArray[0];
-    String result = "";
-      for (int j = 1; j <= referenceWord.length(); j++) {
-        String prefix = referenceWord.substring(0, j);
-        if (isPresentInAllWords(prefix, inputArray) && result.length() < prefix.length()) {
-          result = prefix;
-        }
-      }
-    return result;
-  }
-  
-  private static boolean isPresentInAllWords(String prefix, String[] words) {
-    int n = words.length, k;
-    for (k = 1; k < n; k++) {
-      if (!words[k].startsWith(prefix)) {
-        return false;
-      }
-    }
-    return true;
   }
 
   // Return permutation matrix mapping input names to adaptedFrame colnames
